fix(stripe-subscription): styling and cleanup of unused components
